# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Ticket Booking System",
        "title": "Ticket Booking System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/payments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Создает новую запись о платеже для бронирования. / [EN] Creates a new payment record for a booking.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "[RU] Создать платеж / [EN] Create payment",
                "parameters": [
                    {
                        "description": "[RU] Данные платежа / [EN] Payment data",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.Payment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/payments/booking/{booking_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Возвращает данные платежа, связанного с указанным бронированием. / [EN] Returns payment data associated with the specified booking.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "[RU] Платеж по ID бронирования / [EN] Get payment by booking ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "[RU] ID бронирования / [EN] Booking ID",
                        "name": "booking_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.Payment"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/payments/webhook": {
            "post": {
                "description": "[RU] Конечная точка для приема уведомлений от платежной системы. / [EN] Endpoint for receiving notifications from the payment system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "[RU] Вебхук платежной системы / [EN] Payment system webhook",
                "parameters": [
                    {
                        "description": "[RU] Данные вебхука / [EN] Webhook data",
                        "name": "webhook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.WebhookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/payments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Возвращает данные конкретного платежа. / [EN] Returns specific payment data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "[RU] Получить информацию о платеже / [EN] Get payment info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "[RU] ID платежа / [EN] Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.Payment"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/payments/{id}/process": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Имитирует процесс оплаты (переводит в статус SUCCESS). / [EN] Simulates the payment process (changes status to SUCCESS).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "[RU] Обработать платеж / [EN] Process payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "[RU] ID платежа / [EN] Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.Payment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_bekzat-kamen_booking_system_api_internal_model.CreatePaymentRequest": {
            "type": "object",
            "required": [
                "booking_id",
                "payment_method"
            ],
            "properties": {
                "booking_id": {
                    "type": "string"
                },
                "payment_method": {
                    "enum": [
                        "card",
                        "sbp",
                        "cash"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.PaymentMethod"
                        }
                    ]
                }
            }
        },
...
        "github_com_bekzat-kamen_booking_system_api_internal_model.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "booking_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "paid_at": {
                    "type": "string"
                },
                "payment_method": {
                    "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.PaymentMethod"
                },
                "provider": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.PaymentStatus"
                },
                "transaction_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_bekzat-kamen_booking_system_api_internal_model.PaymentMethod": {
            "type": "string",
            "enum": [
                "card",
                "sbp",
                "cash"
            ],
            "x-enum-varnames": [
                "PaymentMethodCard",
                "PaymentMethodSBP",
                "PaymentMethodCash"
            ]
        },
        "github_com_bekzat-kamen_booking_system_api_internal_model.PaymentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "success",
                "failed",
                "refunded"
            ],
            "x-enum-varnames": [
                "PaymentStatusPending",
                "PaymentStatusSuccess",
                "PaymentStatusFailed",
                "PaymentStatusRefunded"
            ]
        },
...
        "github_com_bekzat-kamen_booking_system_api_internal_model.WebhookRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "signature": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}